fix: auth
