+++ internal/domain/model/response/home.go
@@ -8,3 +8,11 @@ type AllMentor struct {
 	Motto          string `json:"Motto"`
 	ProfilePicture string `json:"ProfilePicture"`
 }
+type TopMentor struct {
+	ID             string `json:"ID"`
+	Name           string `json:"Name"`
+	Type           string `json:"Type"`
+	Description    string `json:"Description"`
+	Motto          string `json:"Motto"`
+	ProfilePicture string `json:"ProfilePicture"`
+}
